Separation of constants: Constants for tables and risk values have been moved to a separate file (Constants.java).
Injection of dynamic dependencies: DBService now receives a map of functions (depsLoader) that represent the interactions with the external.

Updated driver: The driver injects dependencies into DBService using the dynamic depsLoader created in Utils.createDbLoader().

Benefits:
Decoupling: Now the interactions with the database are decoupled and injected dynamically, making it easier to change implementations without affecting the business logic.